+++ src/GafferCortexModule/ParameterisedHolderBinding.cpp
@@ -68,7 +68,7 @@ template<typename T>
 class ParameterisedHolderSerialiser : public NodeSerialiser
 {
 
-	std::string postScript( const Gaffer::GraphComponent *graphComponent, const std::string &identifier, const Serialisation &serialisation ) const override
+	std::string postScript( const Gaffer::GraphComponent *graphComponent, const std::string &identifier, Serialisation &serialisation ) const override
 	{
 		const T *parameterisedHolder = static_cast<const T *>( graphComponent );