+++ src/GafferModule/CompoundNumericPlugBinding.cpp
@@ -87,7 +87,7 @@ class CompoundNumericPlugSerialiser : public ValuePlugSerialiser
 
 	public :
 
-		std::string constructor( const Gaffer::GraphComponent *graphComponent, const Serialisation &serialisation ) const override
+		std::string constructor( const Gaffer::GraphComponent *graphComponent, Serialisation &serialisation ) const override
 		{
 			return serialisationRepr( static_cast<const T *>( graphComponent ), &serialisation );
 		}