+++ src/GafferModule/Transform2DPlugBinding.cpp
@@ -62,7 +62,7 @@ class Transform2DPlugSerialiser : public ValuePlugSerialiser
 			return false;
 		}
 
-		std::string constructor( const Gaffer::GraphComponent *graphComponent, const Serialisation &serialisation ) const override
+		std::string constructor( const Gaffer::GraphComponent *graphComponent, Serialisation &serialisation ) const override
 		{
 			return repr( static_cast<const Transform2DPlug *>( graphComponent ), &serialisation );
 		}